re #56
make methods relevant to file read/write synchronized. This approach should work but is too restrictive, especially it does not allow writing or reading on to different files. However, there is no interference and it should be allowed for perfromance reasons.

What we finally want is that file i/o only synchronizes if the same url/file is affected.