super subtle bug!
scenario:
- fresh app install
- first call is fetchPage()
- second call is fetchPage()
- more calls on fetchPage() or any other

observed:
- second call fails while all others work

what happened?
after fresh app installation CacheCompatManger.cleanUp() triggers a cache clear the first time it is called.
The first fetchPage() was instructed to download from network, no cleanUp() is triggered in this path.
The second fetchPage() finds entry in cach index and instructs to download from cache. Here the cleanUp() is triggered - however between the index lookup and the actual cache lookup! Hence, the file is not found.
Additionally, error handling was wrong at this point (because this case never happened before): an error was thrown though there still was the possiblity to download the page.

fixes:
- call cleanUp() at all entry points of ION calls (i.e. call it at the beginning of fetchPage() and not at the beginning of fetchPageFromCache())
- when no cache file is found for page then try to download it (if possible)