Continue Archive download if a page cache index can't be saved
because the "lastChanged" date can't be determined because the page can't be found in the collection.

Before this change, the archive download stopped in the BCGyou app because there was an invalid URL for a media file and getRequestType() wrongfully deferred this to be a page (the number of segments seemed to fit).
This prevented the Offline-App from a successful login.